refactor(formats): factor out pyarrow and pandas output handling

+++ ibis/backends/tests/snapshots/test_sql/test_isin_bug/duckdb/out.sql
@@ -7,7 +7,7 @@ SELECT
         t0.x AS x
       FROM t AS t0
       WHERE
-        t0.x > CAST(2 AS SMALLINT)
+        t0.x > CAST(2 AS TINYINT)
     ) AS t1
   ) AS "Contains(x, x)"
 FROM t AS t0
